requirements.txt
pyproj
fiona
pygdal==3.2.2.10
geopandas
matplotlib
rasterio
geojson
gdal2tiles
turfpy
shapely
requests
Pillow==9.4.0
weasyprint
CairoSVG